Implements an IMAP client using KIMAP library.
The EmailSessionJob acts as a central session hub for all jobs.
Any job before being started must select a session. This is the
session that all the jobs connect to.

The EmailLoginJob is used to Login to your mailbox. The SelectJob
selects the mailbox's folder -> INBOX

The EmailSearchJob is used to search for only the emails that
are of the past month. More functionality will be added later
when we want updates.

The EmailFetchJob does the actual fetch.

The SingletonFactory is used by all jobs, to connect to the session.

Builds fine.