Adds Update Functionality to the IMAP Client
Uses the IDLE feature (check IMAP RFC), to go into an IDLE state
after fetching few emails. On receiving a signal that the mailbox
folder has changed, it updates the cache by again fetching all the emails.

Note here, that although the Updates work as expected, this solution is
not my final solution.

I intend to fetch ONLY the new emails, not the whole mailbox folder since
that consumes resources. I'm finding a solution that is not so "hackey".

Builds fine and tested with a test gmail id. Had to enable "Login from insecure apps"
feature for the gmail account, in order to test it, since GMail policies recognize
the client as using obsolete security. Will find a way out for that too, but
that will be probably near the end.