Do not fetch emails, when user deletes an email on the server
When the user deletes some emails on his end, we have an update signal from IDLE.
But we don't download new emails as it is not needed.

Whenever the client is run, it FETCHES one email, and that starts it event loop
of selecting/searching/fetching/updating. Even if there are no new emails in the
mailbox, the client will fetch the last received email to start the sort of event cycle.
If the mailBox is empty, the client exits.

Thing I need to do, is to see what would be a graceful way to do the even loop thing,
or merely notify the user that there are no emails in the mailBox, hence exiting.

Builds fine. Tested with gmail.