Created the Daemon class as a central dispatcher of jobs. Refactored all jobs and made them run via this Daemon
This is better than the previous design of classes. The context of the whole job process looks clearer. We don't have to use the Singleton to access the session. We can also, invoke specific methods from the middle, which earlier was just one way to do. For eg, we may simply call searchEmails() after fetch, IDLE etc.